backend/requirements.txt
Flask==2.3.3
requests==2.28.1
python-dotenv==1.0.1